Fix homepage links, scene preview modal init
