--- app-frontend/src/app/components/scenes/sceneDetailModal/sceneDetailModal.component.js
@@ -1,14 +1,0 @@
-import sceneDetailModalTpl from './sceneDetailModal.html';
-
-const rfSceneDetailModal = {
-    templateUrl: sceneDetailModalTpl,
-    bindings: {
-        close: '&',
-        dismiss: '&',
-        modalInstance: '<',
-        resolve: '<'
-    },
-    controller: 'SceneDetailModalController'
-};
-
-export default rfSceneDetailModal;
