+++ app-frontend/src/app/pages/home/home.html
@@ -19,11 +19,12 @@
               <p class="pb-25">
                 Use the {{$ctrl.BUILDCONFIG.APP_NAME}} Tools to build custom processing tasks and get the most out of your imagery.
               </p>
-              <a class="btn btn-secondary" ui-sref="market.search">Browse Tools</a>
+              <a class="btn btn-primary" ui-sref="lab.browse.templates">Browse Templates</a>
+              <a class="btn btn-primary" ui-sref="lab.browse.tools">Browse Tools</a>
               <div class="hide-sm"></div>
               <button class="btn btn-default"
                       ng-click="$ctrl.openToolCreateModal()">
-                Create a new Tool
+                Create a new Template
               </button>
               <!-- <p> -->
               <!-- <a href="https://help.rasterfoundry.com/working-with-tools" target="_blank">Getting started with Tools</a> -->
